# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "mementos-backend",
        "description": "Documentation de l'API du projet Mementos",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/create-user": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Créer un utilisateur",
                "description": "Enregistre un nouvel utilisateur dans la base de données et retourne les informations de l'utilisateur avec un token d'accès.",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Informations nécessaires pour créer un nouvel utilisateur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pseudo",
                                    "email",
                                    "password",
                                    "niveauRevision",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "pseudo": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "niveauRevision": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@exemple.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Pa$$w0rd!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Pa$$w0rd!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inscription réussie"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|RI1TW2GCppDXR09fE0qyVufn8PBZIYeUk6ohC9ruf180affb"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "properties": {
                                                "pseudo": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The pseudo field must be at least 2 characters."
                                                    }
                                                },
                                                "niveauRevision": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The niveau revision field must not be greater than 7."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email has already been taken."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The password field must be at least 8 characters."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur interne au serveur s'est produite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connecter un utilisateur",
                "description": "Permet à un utilisateur de se connecter en utilisant ses credentials et retourne un token d'accès en cas de succès.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Les identifiants de connexion nécessaires",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@exemple.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Pa$$w0rd!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur connecté avec succès"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "6|pnlfhRkVBhmFiJKV1Qi..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur d'authentification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email ou mot de passe incorrect"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur interne au serveur s'est produite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "Schema d'un user",
                "required": [
                    "id",
                    "pseudo",
                    "email",
                    "niveauRevision",
                    "subscribedNotifications",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "pseudo": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01T00:00:00Z"
                    },
                    "niveauRevision": {
                        "type": "integer",
                        "example": 3
                    },
                    "subscribedNotifications": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-02T00:00:00Z"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "YourSecurePassword"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}